backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
slowapi==0.1.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary]==3.1.18
asyncpg==0.29.0

# Redis and Caching
redis==5.0.1
hiredis==2.3.2

# Authentication and Security
PyJWT[crypto]==2.8.0
# GAP-24: passlib is unmaintained (last release Jun 2020). Still functional but
# consider migrating to argon2-cffi or direct bcrypt when convenient.
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1
pydantic-settings==2.1.0

# Background Tasks
celery==5.3.6
flower==2.0.1

# HTTP Client
httpx==0.27.0
socksio>=1.0.0,<2.0.0  # SOCKS5 proxy support for httpx (Tor .onion routing)
aiohttp>=3.11.0  # GAP-15: >=3.10.11 required for CVE-2024-23334 + CVE-2024-23829

# Docker (Dev Sandbox)
docker>=7.0.0

# Nostr
pynostr>=0.7.0
websockets>=12.0

# AI/ML Libraries
openai==1.82.0
anthropic==0.52.0
tiktoken==0.9.0

# Vector Database (optional - commented out due to C++ build requirements)
# chromadb==0.4.22

# Validation and Serialization
pydantic==2.11.4
pydantic[email]==2.11.4
json5>=0.9.0

# Utilities
python-slugify==8.0.4
pytz==2024.1
aiofiles==23.2.1
python-magic==0.4.27
Pillow>=10.4.0  # GAP-15: >=10.3.0 required for CVE-2024-28219

# Testing
pytest>=8.2.0
pytest-asyncio>=0.24.0
pytest-cov>=4.1.0
httpx==0.27.0

# Development
black==24.2.0
ruff==0.2.2
mypy==1.8.0
zhipuai>=2.1.5
aiosqlite>=0.19.0

# Security — ML-based prompt injection detection (optional but recommended)
llm-guard>=0.3.14
